fix: reinstall with the hoisted linker before releasing so npm can walk node_modules
The releases kept failing after the legacy-peer-deps restore: npm's arborist
cannot walk Bun's isolated node_modules layout (the .bun symlink farm) and
crashes with "Cannot read properties of null (reading 'isDescendantOf')" —
the layout #977 switched this repo to, which is when releases started failing.
The release preparation script (renamed from stripWorkspaceProtocol.mjs) now
also flips bunfig.toml to the hoisted linker and reinstalls on the CI checkout
before multi-semantic-release runs. Verified in a worktree: `npm version` and
`npm publish --dry-run` succeed on the hoisted layout and crash on isolated.
